requirements.txt
amqp==2.1.4
appnope==0.1.0
Babel==2.3.4
billiard==3.5.0.2
bleach==2.0.0
celery==4.0.2
cycler==0.10.0
decorator==4.0.11
Django==1.10.5
django-braces==1.11.0
django-oauth-toolkit==0.12.0
djangorestframework==3.5.4
entrypoints==0.2.2
flower==0.9.1
googlefinance==0.7
html5lib==0.999999999
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
kombu==4.0.2
MarkupSafe==1.0
matplotlib==2.0.0
mistune==0.7.4
mpmath==0.19
nbconvert==5.1.1
nbformat==4.3.0
nose==1.3.7
notebook==4.4.1
numpy==1.12.1
oauthlib==2.0.1
pandas==0.19.2
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-magic==0.4.13
pytz==2016.10
pyzmq==16.0.2
qtconsole==4.2.1
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
sympy==1.0
terminado==0.6
testpath==0.3
tornado==4.2
traitlets==4.3.2
ujson==1.35
vine==1.1.3
wcwidth==0.1.7
webencodings==0.5
widgetsnbextension==2.0.0
